Preliminary addition of item_desirability marker to.. Advanced Inventory Manager.
The initial idea was to add a spot of color to the items much like when an item has an auto-pickup entry.  Color turned out to be a tad too much for the first try so opted to add an additional character (1 through 9) to the item line.

The purpose is so that one can track what items you want to gather more of vs those you don't care that much about.  Also handy for highlighting items (tools) that you don't have back at base that your previous character did but because you thought you had one you left that nice thing back in the now too dangerous lab...

I'd like to flesh out this first iteration a bit more before I attempt color again (or if desired I can stick to numbers).

It adds 3 new buttons on the advanced inventory which can be assigned, at the moment they are not assigned by default, nor are they translated.

One key increments the desirability rating (starts at 1, loops back to 1).  Another key decrements the desirability rating (starts at 9, loops back to 9).  The final new key wipes the desirability.

The lines for advanced inventory now look like this: [1-9] [optional symbol character and space][item name][etc...].
When a game is saved, so is the item desirability data.  When a game/character is loaded so is that character's data.

Desirability is stored per character in a new file.

This is the first time I've tried to do something this big in C++ in a VERY long while.