Added a fix for item pickup magenta letter in compact show symbol mode
Without my modifications if you are in compact mode and have show item symbols turned on the first letter of the item name will overwrite the symbol for the item in magenta.  Since I've modified the item layout for now the letter offset is not only added but set to fit my item print.

For example say you had an item line that looked like this:
% Dried Meat ...
and you hit [p] then the line would look like this:
D Dried Meat ...
(with the first D in magenta).  This has been fixed to:
% Dried Meat ...
(with the D in magenta).  If you have item symbols off it will be:
Dried Meat ...
(with the D in magenta). Just like before.