Updated the concept with some research on items and pockets...
Formerly was doing 'string cleaning', now using item labels.  Labels are a cleaner version of the text with no care about charges in magazine or temperature.

Also code is more interested in the item in a container than the container.  For example if the game would show 'plastic bottle > clean water' then the code is more interested in the clean water for item tracking.