Stopped using item::has_label(), using string comparison to 'none' instead...
Did a bunch of testing and whether incompetence or bug I could not get item::has_label() to return true if called from outside the item class.  The method is public but it has different behavior from within or without the class.

Since I knew the no-label string was 'none' used that as gatekeeper on if it should care about item contents for potential interest.